add action for getfriendlist connect api
